+++ packages/berlin/src/layout/Layout.tsx
@@ -8,13 +8,12 @@ import { Main } from './Layout.styled';
 // Components
 // import Header from '../components/header';
 import Footer from '../components/footer';
-import NewHeader from '@/components/header/NewHeader';
+import NewHeader from '@/components/header';
 
 function Layout() {
   return (
     <>
       <Toaster position="top-center" toastOptions={{ style: { fontFamily: 'Raleway' } }} />
-      {/* <Header /> */}
       <NewHeader />
       <Main>
         <Outlet />
